+++ mq/main/bridge/bridge-jms/src/main/java/com/sun/messaging/bridge/service/jms/SharedConnectionFactory.java
@@ -100,14 +100,14 @@ public class SharedConnectionFactory implements Runnable {
         } 
         _idleTimeout = Integer.valueOf(attrs.getProperty(
                                        JMSBridgeXMLConstant.CF.IDLETIMEOUT, 
-                                       JMSBridgeXMLConstant.CF.IDLETIMEOUT_DEFAULT));
+                                       JMSBridgeXMLConstant.CF.IDLETIMEOUT_DEFAULT)).intValue();
         if (_idleTimeout < 0) _idleTimeout = 0;
         _maxRetries = Integer.valueOf(attrs.getProperty(
                                       JMSBridgeXMLConstant.CF.CONNECTATTEMPTS,
-                                      JMSBridgeXMLConstant.CF.CONNECTATTEMPTS_DEFAULT));
+                                      JMSBridgeXMLConstant.CF.CONNECTATTEMPTS_DEFAULT)).intValue();
         _retryInterval = Integer.valueOf(attrs.getProperty(
                                          JMSBridgeXMLConstant.CF.CONNECTATTEMPTINTERVAL,
-                                         JMSBridgeXMLConstant.CF.CONNECTATTEMPTINTERVAL_DEFAULT));
+                                         JMSBridgeXMLConstant.CF.CONNECTATTEMPTINTERVAL_DEFAULT)).intValue();
         if (_retryInterval < 0) _retryInterval = 0;
 
         _scheduler = Executors.newSingleThreadScheduledExecutor();
